priority double

+++ src/dataStructures/PriorityNode.java
@@ -2,14 +2,14 @@ package dataStructures;
 
 public class PriorityNode<E> {
 	public E key;
-	public int priority;
+	public double priority;
     public E getKey() {
 		return key;
 	}
 	public void setKey(E key) {
 		this.key = key;
 	}
-	public int getPriority() {
+	public double getPriority() {
 		return priority;
 	}
 	public void setPriority(int priority) {
@@ -18,7 +18,7 @@ public class PriorityNode<E> {
 	protected PriorityNode<E> next;
     protected PriorityNode<E> prev;
     
-    public PriorityNode(E key,int priority,PriorityNode<E> next,PriorityNode<E> prev){
+    public PriorityNode(E key,double priority,PriorityNode<E> next,PriorityNode<E> prev){
         this.key = key;
         this.next = next;
         this.prev = prev;
